adding Surface component for displaying Paper vis
