+++ client/src/components/Surface/canvasStyle.js
@@ -1,0 +1,6 @@
+export default  {
+  width: '400px',
+  height: '200px',
+  border: '1px solid #eee',
+  backgroundColor: 'white'
+};
